backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
openai>=1.69.0
pydantic==2.9.0
arize-otel
openinference-instrumentation-openai
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
python-dotenv==1.0.1
pytest>=8.0
pytest-asyncio>=0.23
